@@ -22,41 +22,40 @@
   "dependencies": {
     "jile": "^2.0.2",
     "lodash": "^4.17.2",
-    "react": "^15.4.1",
     "uuid": "^3.0.1"
   },
   "description": "Decorator for jile usage on React components",
   "devDependencies": {
-    "ava": "^0.17.0",
-    "babel": "^6.5.2",
+    "ava": "^0.25.0",
     "babel-cli": "^6.18.0",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.2.8",
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^8.2.5",
+    "babel-loader": "^7.1.5",
     "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-preset-latest": "^6.16.0",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-2": "^6.18.0",
-    "enzyme": "^2.6.0",
-    "eslint": "^3.11.1",
-    "eslint-config-rapid7": "^2.6.0",
-    "eslint-friendly-formatter": "^2.0.6",
-    "eslint-loader": "^1.6.1",
-    "html-webpack-plugin": "^2.24.1",
+    "browser-env": "^3.2.5",
+    "enzyme": "^3.8.0",
+    "enzyme-adapter-react-16": "^1.7.1",
+    "eslint": "^5.12.0",
+    "eslint-config-rapid7": "^3.1.0",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-loader": "^2.1.1",
+    "html-webpack-plugin": "^3.2.0",
     "in-publish": "^2.0.0",
-    "jsdom": "^9.8.3",
-    "jsdom-global": "^2.1.0",
-    "lodash-webpack-plugin": "^0.10.6",
-    "nyc": "^10.0.0",
-    "onchange": "^3.2.0",
+    "nyc": "^13.1.0",
+    "onchange": "^5.2.0",
     "optimize-js-plugin": "^0.0.4",
+    "prop-types": "^15.6.2",
+    "react": "^16.7.0",
     "react-addons-test-utils": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "react-dom": "^16.7.0",
     "rimraf": "^2.5.4",
-    "sinon": "^1.17.6",
-    "webpack": "^1.13.3",
-    "webpack-dashboard": "^0.2.0",
-    "webpack-dev-server": "^1.16.2"
+    "sinon": "^7.2.2",
+    "webpack": "^4.28.4",
+    "webpack-cli": "^3.2.1",
+    "webpack-dev-server": "^3.1.14"
   },
   "homepage": "https://github.com/planttheidea/react-jile#readme",
   "keywords": [
@@ -68,6 +67,9 @@
   "license": "MIT",
   "main": "lib/index.js",
   "name": "react-jile",
+  "peerDependencies": {
+    "react": "^15.3.0 || ^16.0.0"
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/planttheidea/react-jile.git"
@@ -76,7 +78,7 @@
     "build": "NODE_ENV=development webpack --progress --colors",
     "build:minified": "NODE_ENV=production webpack --progress --colors --config=webpack.config.minified.js",
     "clean": "rimraf lib && rimraf dist",
-    "dev": "NODE_ENV=development webpack-dashboard -- webpack-dev-server --config=webpack.config.dev.js",
+    "dev": "NODE_ENV=development webpack-dev-server --progress --config=webpack.config.dev.js",
     "lint": "NODE_ENV=test eslint src",
     "prepublish": "in-publish && npm run prepublish:compile || echo ''",
     "prepublish:compile": "npm run clean && npm run lint && npm test && npm run transpile && npm run build && npm run build:minified",
