@@ -2,58 +2,37 @@
 
 const path = require('path');
 const webpack = require('webpack');
-const LodashModuleReplacementPlugin = require('lodash-webpack-plugin');
 
 module.exports = {
   devtool: '#source-map',
 
-  entry: [
-    path.resolve (__dirname, 'src', 'index.js')
-  ],
+  entry: [path.resolve(__dirname, 'src', 'index.js')],
 
-  eslint: {
-    configFile: '.eslintrc',
-    emitError: true,
-    failOnError: true,
-    failOnWarning: true,
-    formatter: require('eslint-friendly-formatter')
-  },
+  externals: ['hash-it', 'jile', 'react'],
 
-  externals: {
-    'jile': {
-      amd: 'jile',
-      commonjs: 'jile',
-      commonjs2: 'jile',
-      root: 'jile'
-    },
-    'react': {
-      amd: 'react',
-      commonjs: 'react',
-      commonjs2: 'react',
-      root: 'React'
-    }
-  },
+  mode: 'development',
 
   module: {
-    preLoaders: [
+    rules: [
       {
-        include: [
-          /src/
-        ],
+        enforce: 'pre',
+        include: [path.join(__dirname, 'src')],
         loader: 'eslint-loader',
-        test: /\.js$/
-      }
-    ],
-
-    loaders: [
+        options: {
+          configFile: '.eslintrc',
+          emitError: true,
+          failOnError: true,
+          failOnWarning: false,
+          formatter: require('eslint-friendly-formatter'),
+        },
+        test: /\.js$/,
+      },
       {
-        include: [
-          /src/
-        ],
-        loader: 'babel',
-        test: /\.js?$/
-      }
-    ]
+        include: [path.join(__dirname, 'src'), path.join(__dirname, 'DEV_ONLY')],
+        loader: 'babel-loader',
+        test: /\.js?$/,
+      },
+    ],
   },
 
   output: {
@@ -61,24 +40,8 @@ module.exports = {
     library: 'ReactJile',
     libraryTarget: 'umd',
     path: path.resolve(__dirname, 'dist'),
-    umdNamedDefine: true
+    umdNamedDefine: true,
   },
 
-  plugins: [
-    new webpack.EnvironmentPlugin([
-      'NODE_ENV'
-    ]),
-    new LodashModuleReplacementPlugin({
-      collections: true
-    })
-  ],
-
-  resolve: {
-    extensions: [
-      '',
-      '.js'
-    ],
-
-    root: __dirname
-  }
+  plugins: [new webpack.EnvironmentPlugin(['NODE_ENV'])],
 };
