@@ -1,36 +1,22 @@
 'use strict';
 
-const webpack = require("webpack");
+const webpack = require('webpack');
 const OptimizeJsPlugin = require('optimize-js-plugin');
 
-const defaultConfig = require("./webpack.config");
+const defaultConfig = require('./webpack.config');
 
 module.exports = Object.assign({}, defaultConfig, {
-  devtool: null,
+  devtool: false,
+
+  mode: 'production',
 
   output: Object.assign({}, defaultConfig.output, {
-    filename:"react-jile.min.js"
+    filename: 'react-jile.min.js',
   }),
 
   plugins: defaultConfig.plugins.concat([
-    new webpack.optimize.DedupePlugin(),
-    new webpack.optimize.OccurrenceOrderPlugin(),
-    new webpack.optimize.UglifyJsPlugin({
-      compress:{
-        booleans:true,
-        conditionals:true,
-        drop_console:false,
-        drop_debugger:true,
-        join_vars:true,
-        screw_ie8:true,
-        sequences:true,
-        warnings:false
-      },
-      mangle:true,
-      sourceMap:false
-    }),
     new OptimizeJsPlugin({
-      sourceMap: false
-    })
-  ])
+      sourceMap: false,
+    }),
+  ]),
 });
